+++ src/ethpector/symbolic/modules/extractor.py
@@ -2,6 +2,10 @@ import logging
 from pyevmasm import instruction_tables, DEFAULT_FORK
 from mythril.analysis.module.base import DetectionModule, EntryPoint
 from mythril.laser.ethereum.state.global_state import GlobalState
+from mythril.exceptions import UnsatError
+from mythril.laser.smt import UGT, symbol_factory
+from mythril.analysis import solver
+from copy import copy
 
 # from mythril.laser.ethereum.transaction.transaction_models import (
 #     ContractCreationTransaction,
@@ -54,6 +58,8 @@ class RecoverData(DetectionModule):
         ]  # GETPC currently not supported by mythril
         super().__init__()
         self.public_functions = {}
+        self.reach_with_value = set()
+        self.revert_with_value = set()
         self.calls = []
         self.storage_reads = []
         self.storage_writes = []
@@ -71,6 +77,11 @@ class RecoverData(DetectionModule):
         self.creates = []
         self.create2s = []
         self.seen_pcs = set()
+
+    def payable(self):
+        return (
+            set(self.public_functions.keys()) - self.revert_with_value
+        ) - self.reach_with_value
 
     def get_annotations(self):
         return (
@@ -94,6 +105,18 @@ class RecoverData(DetectionModule):
 
     # def reset_module(self):
     #     super().reset_module()
+
+    def can_reach_with_value(self, state: GlobalState) -> bool:
+        op, pc, instruction, stack, func, _ = decompose_inst(state)
+        constraints = copy(state.world_state.constraints)
+        constraints.append(
+            UGT(state.current_transaction.call_value, symbol_factory.BitVecVal(0, 256))
+        )
+        try:
+            solver.get_model(constraints)
+            return True
+        except UnsatError:
+            return False
 
     def _execute(self, state: GlobalState) -> None:
         if state.get_current_instruction()["address"] in self.cache:
@@ -121,12 +144,19 @@ class RecoverData(DetectionModule):
         self.public_functions[func].update(state)
 
         if op in CALL_INSTRUCTION_LIST:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
+
             self.calls.append(Call.from_statespace(state))
 
         if op in ["SSTORE"]:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.storage_writes.append(StorageWrite.from_statespace(state))
 
         if op in ["SLOAD"]:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.storage_reads.append(StorageLoad.from_statespace(state))
 
         if op in ["MSTORE"]:
@@ -136,18 +166,26 @@ class RecoverData(DetectionModule):
             self.memory_reads.append(MemoryLoad.from_statespace(state))
 
         if op in ["RETURN"]:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.returns.append(Return.from_statespace(state))
 
         if op in ["REVERT"]:
+            if self.can_reach_with_value(state):
+                self.revert_with_value.add(func)
             self.reverts.append(Revert.from_statespace(state))
 
         if op in ["SELFDESTRUCT"]:
             self.selfdestructs.append(Selfdestruct.from_statespace(state))
 
         if op in ["CALLDATALOAD"]:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.calldataloads.append(Calldataload.from_statespace(state))
 
         if op in ["CALLDATACOPY"]:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.calldatacopies.append(Calldatacopy.from_statespace(state))
 
         if op in ["JUMP"]:
@@ -160,12 +198,18 @@ class RecoverData(DetectionModule):
             self.pushes.append(Push.from_statespace(state))
 
         if "CREATE2" == op:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.create2s.append(Create2.from_statespace(state))
 
         if "CREATE" == op:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.creates.append(Create.from_statespace(state))
 
         if op in LOG_INSTRUCTION_LIST:
+            if self.can_reach_with_value(state):
+                self.reach_with_value.add(func)
             self.logs.append(Log.from_statespace(state))
 
         return []